requirements.txt
# from the google quickstart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
